pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-screenshot-server"
version = "0.1.0"
description = "MCP server for screenshot capture and image annotation with boxes, lines, arrows, and text"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MCP Screenshot Server Contributors" }
]
keywords = ["mcp", "screenshot", "annotation", "image-processing", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
clipboard = [
    "pyperclip>=1.8.0",
]

[project.scripts]
mcp-screenshot-server = "mcp_screenshot_server.server:main"

[project.urls]
Homepage = "https://github.com/aamar-shahzad/mcp-screenshot-server"
Repository = "https://github.com/aamar-shahzad/mcp-screenshot-server"
Issues = "https://github.com/aamar-shahzad/mcp-screenshot-server/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_screenshot_server"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = ["E501"]  # Line length handled separately

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

